rafactor: Refactoring of backend package architecture
