+++ src/main/java/com/example/kinnibackend/repository/example/TestRepository.java
@@ -1,6 +1,6 @@
-package com.example.kinnibackend.example.repository;
+package com.example.kinnibackend.repository.example;
 
-import com.example.kinnibackend.example.entity.Test;
+import com.example.kinnibackend.entity.Test;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TestRepository extends JpaRepository<Test, Long> {
