+++ src/main/java/com/example/kinnibackend/service/example/TestService.java
@@ -1,9 +1,9 @@
-package com.example.kinnibackend.example.service;
+package com.example.kinnibackend.service.example;
 
-import com.example.kinnibackend.example.dto.CreateTestRequestDTO;
-import com.example.kinnibackend.example.dto.TestResponseDTO;
-import com.example.kinnibackend.example.entity.Test;
-import com.example.kinnibackend.example.repository.TestRepository;
+import com.example.kinnibackend.dto.example.CreateTestRequestDTO;
+import com.example.kinnibackend.dto.example.TestResponseDTO;
+import com.example.kinnibackend.entity.Test;
+import com.example.kinnibackend.repository.example.TestRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
